Refactor: 미사용 모듈 정리 및 디렉토리 구조 변경.
