+++ src/advanced/components/admin/CouponList.tsx
@@ -6,6 +6,7 @@ import {
   pushNotificationAtom,
   showCouponFormAtom,
 } from '../../store/atoms';
+import { formatPrice } from '../../utils/formatters';
 
 export default function CouponList() {
   const [coupons] = useAtom(couponsAtom);
@@ -35,7 +36,7 @@ export default function CouponList() {
               <div className='mt-2'>
                 <span className='inline-flex items-center px-3 py-1 rounded-full text-sm font-medium bg-white text-indigo-700'>
                   {coupon.discountType === 'amount'
-                    ? `${coupon.discountValue.toLocaleString()}원 할인`
+                    ? `${formatPrice(coupon.discountValue)} 할인`
                     : `${coupon.discountValue}% 할인`}
                 </span>
               </div>
