--- src/advanced/hooks/useCart.ts
@@ -1,144 +1,0 @@
-// TODO: 장바구니 관리 Hook
-// 힌트:
-// 1. 장바구니 상태 관리 (localStorage 연동)
-// 2. 상품 추가/삭제/수량 변경
-// 3. 쿠폰 적용
-// 4. 총액 계산
-// 5. 재고 확인
-//
-// 사용할 모델 함수:
-// - cartModel.addItemToCart
-// - cartModel.removeItemFromCart
-// - cartModel.updateCartItemQuantity
-// - cartModel.calculateCartTotal
-// - cartModel.getRemainingStock
-//
-// 반환할 값:
-// - cart: 장바구니 아이템 배열
-// - selectedCoupon: 선택된 쿠폰
-// - addToCart: 상품 추가 함수
-// - removeFromCart: 상품 제거 함수
-// - updateQuantity: 수량 변경 함수
-// - applyCoupon: 쿠폰 적용 함수
-// - calculateTotal: 총액 계산 함수
-// - getRemainingStock: 재고 확인 함수
-// - clearCart: 장바구니 비우기 함수
-
-import { useEffect, useState } from 'react';
-
-import { CartItem, Coupon, Product, ProductWithUI } from '../../types';
-
-export function useCart() {
-  const [cart, setCart] = useState<CartItem[]>(() => {
-    const saved = localStorage.getItem('cart');
-    if (saved) {
-      try {
-        return JSON.parse(saved);
-      } catch {
-        return [];
-      }
-    }
-    return [];
-  });
-
-  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
-
-  useEffect(() => {
-    if (cart.length > 0) {
-      localStorage.setItem('cart', JSON.stringify(cart));
-    } else {
-      localStorage.removeItem('cart');
-    }
-  }, [cart]);
-
-  const addToCart = (product: ProductWithUI, onSuccess?: () => void) => {
-    const remainingStock = getRemainingStock(product);
-
-    let error = null;
-
-    if (remainingStock <= 0) {
-      error = new Error(`상품 "${product.name}"의 재고가 부족합니다.`);
-      return;
-    }
-
-    setCart((prevCart) => {
-      const existingItem = prevCart.find((item) => item.product.id === product.id);
-
-      if (existingItem) {
-        const newQuantity = existingItem.quantity + 1;
-
-        if (newQuantity > product.stock) {
-          error = new Error(`상품 "${product.name}"의 재고는 ${product.stock}개까지만 가능합니다.`);
-          return prevCart;
-        }
-
-        return prevCart.map((item) =>
-          item.product.id === product.id ? { ...item, quantity: newQuantity } : item,
-        );
-      }
-
-      return [...prevCart, { product, quantity: 1 }];
-    });
-
-    if (error) throw error;
-
-    onSuccess?.();
-  };
-
-  const removeFromCart = (productId: string) => {
-    setCart((prevCart) => prevCart.filter((item) => item.product.id !== productId));
-  };
-
-  const updateQuantity = (productId: string, newQuantity: number) => {
-    setCart((prevCart) =>
-      prevCart.map((item) =>
-        item.product.id === productId ? { ...item, quantity: newQuantity } : item,
-      ),
-    );
-  };
-
-  const applyCoupon = (coupon: Coupon | null) => {
-    setSelectedCoupon(coupon);
-  };
-
-  const calculateTotal = () => {
-    let total = 0;
-    cart.forEach((item) => {
-      total += item.product.price * item.quantity;
-    });
-
-    if (selectedCoupon) {
-      if (selectedCoupon.discountType === 'amount') {
-        total -= selectedCoupon.discountValue;
-      } else {
-        total *= 1 - selectedCoupon.discountValue / 100;
-      }
-    }
-
-    return Math.max(total, 0); // 총액은 0 이상
-  };
-
-  const getRemainingStock = (product: Product): number => {
-    const cartItem = cart.find((item) => item.product.id === product.id);
-    const remaining = product.stock - (cartItem?.quantity || 0);
-
-    return remaining;
-  };
-
-  const clearCart = () => {
-    setCart([]);
-    setSelectedCoupon(null);
-  };
-
-  return {
-    cart,
-    selectedCoupon,
-    addToCart,
-    removeFromCart,
-    updateQuantity,
-    applyCoupon,
-    calculateTotal,
-    getRemainingStock,
-    clearCart,
-  };
-}
